Use correct Logic App trigger url in logic_app_receiver
- In commit fca604ae5ca185de5819e6d1f2432f57d4f912cf a change to the
logic_app_receiver block in the monitor action group changed the trigger
url of the Logic App Workflow to be dynamic based on whether the operator
is supplying their own Logic App workflow resource or choosing to create
one with the module.
- If operators chose to create the Logic app workflow using the
module, the actual value that was set was invalid and only contained the
base url of the workflow. This meant that any requests
to the workflow would fail silenty.
- The trigger url must contain an authenticated token which is only
possible if using the callback_url attribute of the HTTP Trigger.
- The callback_url attribute cannot be gotten from a data source so the fix is
to load it using a AzApi request instead.

 # Please enter the commit message for your changes.
Lines starting